requirements.txt
streamlit
replicate
python-dotenv
toml
whois
python-docx